Now, if player has 0 amount of rangers, the mission will fail
In the original triggers, if you can bring at least one white archer to circle of power, the player wins. But if player loses all white archers, the game doesn't end, and there's no way player can win. Therefore the mission gets stuck. This change could be considered a kind of finetuning that makes the mission automatically lost if all white archers are lost.
human/level02h.sms file (the file that determines the map) should also be edited in commemoration and white archers should be replaced with white rangers for these triggers to work.