Victory condition triggers and define allow fixed
In the original, the mission objectives were "-Destroy all strongholds and fortresses","-Turalyon must reach the Dark Portal alive". But the former triggers dictated that you must eliminate all enemy units in order to win, not just strongholds and fortresses. Also added Turalyon reaching the circle of power condition as the former triggers didn't have that. And added all paladin spells to be readily available from start, as that was how it was for this mission in the original.