Polished defeat conditions
Added condition that if the captured human mage dies, you automatically lose the scenario. Since, you are expected to bring him alive to the circle of power in order to win the scenario. In the case of mage's dead, the scenario used to be in a deadlock where it was impossible for you to win the game, as such it must mean that you've automatically lost the game.