upgrade.lua (my changes for gnimish fighter upgrade)
You can see that in lines 85 and 86, I've initialized that upgrade. In lines 175 and 176, I've declared that the upgrade does the convertion. In lines 364 and 365, I've defined dependencies for that upgrade.
Did I have to make changes in the code block that's between lines 89 and 95 for this to work? Because I don't understand what that block of lines says, and I couldn't just make a modification there without understanding.