new spells.lua update
Demolition: range 2 and damage 400 is way too OP and isn't as how it was in original warcraft. I made it so it's range 1 and damage 400, and an additional range 2 and damage 0. The second line of code will make it so that it destroys trees and rocks 2 tiles away just like how it used to in warcraft 2, but it won't deal damage to units.

Holy vision: 25 is too short of a time to recognize anything on the map after casting the spell. I made it 100 so that it lasts a little bit longer (not too long), so that you can actually see what's going on after casting the spell, as how it was in original warcraft 2.